+++ apps/api/src/issues/activity.ts
@@ -11,7 +11,7 @@ import {
 } from '@repo/db';
 import { and, desc, eq, gte, inArray, isNull, lt, sql } from 'drizzle-orm';
 import { HttpError, iso } from '../shared/lib';
-import { emitWebhookEvent } from '../webhooks/emit';
+import { emitWebhookEvent } from '#modules/webhooks/emit';
 import { parseMentionHandles, resolveMentionHandles } from '#shared/mentions';
 import { isAgentUser, listMentionTriggerAgents } from '#modules/agents/core/service';
 import { enqueueAgentRun } from '#modules/agents/core/run-queue';
