+++ apps/api/src/issues/links.ts
@@ -3,7 +3,7 @@ import { and, eq, isNull, or, sql } from 'drizzle-orm';
 import { alias } from 'drizzle-orm/pg-core';
 import { HttpError } from '../shared/lib';
 import { recordActivityEntries, type ActivityInput } from './activity';
-import { emitWebhookEvents } from '../webhooks/emit';
+import { emitWebhookEvents } from '#modules/webhooks/emit';
 import { getIssues } from './store';
 
 // Relations between issues of one project. A relation is one row (issue_link);
