+++ apps/api/src/modules/webhooks/__tests__/integration/webhooks.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach } from 'bun:test';
-import { authedApi, type Api } from '../../../__tests__/helpers/app';
-import { signUpTestUser } from '../../../__tests__/helpers/auth';
-import { resetDb } from '../../../__tests__/helpers/db';
-import type { WebhookEventType } from '../../store';
+import { authedApi, type Api } from '#tests/helpers/app';
+import { signUpTestUser } from '#tests/helpers/auth';
+import { resetDb } from '#tests/helpers/db';
+import type { WebhookEventType } from '../../service';
 
 async function setupOwnerProject() {
   const owner = await signUpTestUser();
