+++ apps/api/src/modules/webhooks/emit.ts
@@ -1,13 +1,13 @@
 import { randomUUID } from 'node:crypto';
 import { db, webhook, webhookDelivery } from '@repo/db';
 import { and, eq, sql } from 'drizzle-orm';
-import type { WebhookEventType } from './store';
+import type { WebhookEventType } from './service';
 
 // Maps our granular event type to the Linear-style envelope's action + resource
-// type. action is create | update | remove; type is the PascalCase resource.
-// Linear's action/type pair cannot distinguish our assigned / state_changed /
-// label_changed variants (all of them are an issue update), so the granular event
-// is kept in the payload's `event` field.
+// type. action is create | update | remove. type is the PascalCase resource.
+// Linear's action/type pair cannot distinguish our assigned, state_changed, and
+// label_changed variants, because all of them are an issue update. The payload keeps
+// the granular event in its own `event` field.
 const EVENT_SHAPE: Record<WebhookEventType, { action: string; type: string }> = {
   'issue.created': { action: 'create', type: 'Issue' },
   'issue.updated': { action: 'update', type: 'Issue' },
@@ -19,18 +19,18 @@ const EVENT_SHAPE: Record<WebhookEventType, { action: string; type: string }> = 
   'comment.created': { action: 'create', type: 'Comment' },
 };
 
-// Fan-out for outgoing webhooks: queues one delivery per active webhook of the
-// project that is subscribed to eventType, all sharing a single eventId (stable
-// across retries so receivers can deduplicate). Called right after a domain
-// mutation, next to the activity log, following the same post-write side-effect
-// pattern the issue store already uses. No-op when no webhook matches, so it costs
-// one indexed SELECT on projects with no webhooks.
+// Fan-out for outgoing webhooks. Queues one delivery per active webhook of the
+// project that subscribes to eventType. The deliveries share one eventId, which stays
+// the same across retries so a receiver can deduplicate. Call it right after a domain
+// mutation, next to the activity log, the same way the issue store handles its other
+// post-write side effects. It does nothing when no webhook matches, so a project with
+// no webhooks pays one indexed SELECT.
 //
 // The body follows Linear's webhook envelope: top-level action, type, data, plus
-// createdAt and webhookTimestamp (epoch ms). `event` is our extension carrying the
-// granular event type. organizationId and url are omitted (there is no organization
-// concept, and issues have no public URL). The shared dedup id rides on the wire in
-// the X-Itsaplan-Event-Id header, not the body.
+// createdAt and webhookTimestamp (epoch ms). `event` is our extension, and it carries
+// the granular event type. We omit organizationId and url, because there is no
+// organization concept and issues have no public URL. The shared dedup id goes in the
+// X-Itsaplan-Event-Id header, not in the body.
 export async function emitWebhookEvent(
   projectId: number,
   eventType: WebhookEventType,
@@ -39,11 +39,11 @@ export async function emitWebhookEvent(
   await emitWebhookEvents(projectId, eventType, () => Promise.resolve([data]));
 }
 
-// Several events of one type at once, with the payloads assembled only once a
-// webhook turns out to be subscribed. For the writes whose payload costs its own
-// queries to build — linking two issues has to load both of them — so a project
-// with no webhook pays one indexed SELECT and nothing else. Each event gets its
-// own eventId; the deliveries go in as one insert.
+// Several events of one type at once. It builds the payloads only after it finds a
+// subscribed webhook. Use it for a write whose payload costs its own queries to
+// build: linking two issues has to load both of them. A project with no webhook then
+// pays one indexed SELECT and nothing else. Each event gets its own eventId. The
+// deliveries go in as one insert.
 export async function emitWebhookEvents(
   projectId: number,
   eventType: WebhookEventType,
@@ -56,7 +56,7 @@ export async function emitWebhookEvents(
       and(
         eq(webhook.projectId, projectId),
         eq(webhook.isActive, true),
-        // events is a jsonb array of event-type strings; @> tests membership.
+        // events is a jsonb array of event-type strings. @> tests membership.
         sql`${webhook.events} @> ${JSON.stringify([eventType])}::jsonb`,
       ),
     );
