+++ apps/api/src/planner.ts
@@ -21,7 +21,7 @@ import { avatarRoutes } from './modules/avatars';
 import { viewRoutes } from './modules/views';
 import { shareRoutes } from './modules/share';
 import { actionRoutes } from './modules/actions';
-import { webhookRoutes } from './webhooks/routes';
+import { webhookRoutes } from './modules/webhooks';
 import { gitSettingsRoutes } from './modules/git';
 import { dashboardRoutes } from './modules/dashboards';
 import { noteBoardRoutes } from './modules/note-boards';
